+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - No screenings in York for 'basement events'
 - more flexible 2d removal
+- better title sanitizations
 
 ## v.2.0.1 -2014-10-19